Fixed problems with classification algorithms. #746 (CEDMA server is not compatible with new data-modeling algorithms)
git-svn-id: https://src.heuristiclab.com/svn/core/trunk@2363 2abd9481-f8db-48e9-bd25-06bc13291c1b

+++ sources/HeuristicLab.GP.StructureIdentification/3.3/DefaultStructureIdentificationOperators.cs
@@ -155,7 +155,7 @@ namespace HeuristicLab.GP.StructureIdentification {
     }
 
     public static void PopulateAnalyzerModel(IScope bestModelScope, IAnalyzerModel model) {
-      model.SetMetaData("EvaluatedSolutions", bestModelScope.GetVariableValue<DoubleData>("EvaluatedSolutions", false).Data);
+      model.SetMetaData("EvaluatedSolutions", bestModelScope.GetVariableValue<IntData>("EvaluatedSolutions", false).Data);
       IGeneticProgrammingModel gpModel = bestModelScope.GetVariableValue<IGeneticProgrammingModel>("FunctionTree", false);
       model.SetMetaData("TreeSize", gpModel.Size);
       model.SetMetaData("TreeHeight", gpModel.Height);
